refactor(sidebar): fold panelsidebar into rightsidebar
