+++ src/app/server-ai-stream-tool-chatbot-tracked-changes/page.tsx
@@ -12,9 +12,8 @@ import { TrackedChanges } from "@tiptap-pro/extension-tracked-changes";
 import { DefaultChatTransport } from "ai";
 import { type FormEvent, useRef, useState } from "react";
 import { createPortal } from "react-dom";
+import { type PanelId, RightSidebar } from "@/components/right-sidebar";
 import { SuggestionReviewTooltip } from "@/components/suggestion-review-tooltip";
-import type { PanelId } from "@/demos/server-ai-tracked-changes/panel-id";
-import { RightSidebar } from "@/demos/server-ai-tracked-changes/right-sidebar";
 import { TrackedChangesPanel } from "@/demos/server-ai-tracked-changes/tracked-changes-panel";
 import { useStreamToolEditor } from "@/lib/use-stream-tool-editor";
 import { useSuggestionReview } from "@/lib/use-suggestion-review";
