--- src/demos/server-ai-tracked-changes/panel-id.ts
@@ -1,1 +1,0 @@
-export type PanelId = "chat" | "tracked" | "comments";
