+++ src/demos/server-ai-tracked-changes/server-ai-tracked-changes-demo.tsx
@@ -27,10 +27,9 @@ import { v4 as uuid } from "uuid";
 import * as Y from "yjs";
 import { getCollabConfig } from "@/app/server-ai-agent-chatbot/actions";
 import { CopyTestCaseButton } from "@/components/capture-test-case/copy-test-case-button";
+import { type PanelId, RightSidebar } from "@/components/right-sidebar";
 import { SuggestionReviewTooltip } from "@/components/suggestion-review-tooltip";
 import { CommentsPanel } from "./comments-panel";
-import type { PanelId } from "./panel-id";
-import { RightSidebar } from "./right-sidebar";
 import "../../styles/tracked-changes.css";
 import "./server-ai-tracked-changes.css";
 import { getUniqueSuggestions } from "./suggestion-utils";
